requirements.txt
flask==3.0.3
flask-wtf==1.2.1
email-validator==2.2.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
